feat: Update player speed to be affected by speed items
스피드 아이템의 효과를 받을 수 있도록 플레이어 속도 설정을 변경함.
아이템을 획득할 때 플레이어 속도가 증가하도록 수정.

+++ Assets/Scripts/Stats/PlayerStatHandler.cs
@@ -19,12 +19,12 @@ public class PlayerStatHandler : MonoBehaviour
     private void UpdateCharacterStat() // 스탯 업데이트
     {
         PlayerAttackSO playerSO = null;
-        if (baseStats.playerSO != null) // PlayerStat 초기화
+        if (baseStats.playerSO != null) // 공격 스탯 초기화
         {
             playerSO = Instantiate(baseStats.playerSO);
         }
 
-        // 현재 Stat
+        // 현재 플레이어 Stat
         CurrentStat = new PlayerStat { playerSO = playerSO };
         CurrentStat.maxHealth = baseStats.maxHealth;
         CurrentStat.speed = baseStats.speed;
@@ -33,7 +33,7 @@ public class PlayerStatHandler : MonoBehaviour
 
 
     // 스피드 2배로 증가
-    public void ChangeSpeedStat(int multiple)
+    public void ChangeSpeedStat(float multiple)  
     {
         CurrentStat.speed *= multiple;
     }
